requirements.txt
Django==4.0.2
django-debug-toolbar==3.2.4
django-seed==0.3.1
redis==4.1.3
gunicorn==20.1.0
bleach==4.1.0
django-hCaptcha==0.2.0
hashids==1.3.1
django-on-heroku==1.1.2
django_csp==3.7
sentry_sdk==1.5.5
